#88 - Move list test instances into models/instances.js.
